KryptosED/requirements.txt
tkinter
customtkinter
pillow
PyCryptodome